+++ CHANGELOG.md
@@ -20,6 +20,9 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 -   Update `ConversationsService::sendConversationByEmail` with parameters to send it to custom email, and set the language.
 -   Update `ConversationsService::ackAllMessages` with parameter `maskRead` if true Im won't be shown as read on peer conversation side.
 -   Update low layer `RESTService::getServerConversations` method's parameters.
+-   Add `ConversationsService::updateConversationBookmark` API can be used to set or replace a bookmarked message in a conversation.
+-   Add `ConversationsService::loadConversationHistory` API to retrieve the remote history of a specific conversation.
+-   Update the `ConversationsService::getContactsMessagesFromConversationId` method to retrieve history from server if not yet loaded.   
 
 ## [2.20.0] - 2023-02-24
 -   Add `PresenceService::setApplyMsTeamsPresenceSettings` This api allows to activate the exchange of presence of the connected user between rainbow and MS Teams on UI side.